added new dirs and moved files
