--- RunEdgeConvert.java
@@ -1,5 +1,0 @@
-public class RunEdgeConvert {
-   public static void main(String[] args) {
-      EdgeConvertGUI edge = new EdgeConvertGUI();
-   }
-}
