+++ src/main/RunEdgeConvert.java
@@ -1,0 +1,5 @@
+public class RunEdgeConvert {
+   public static void main(String[] args) {
+      EdgeConvertGUI edge = new EdgeConvertGUI();
+   }
+}
