+++ src/test/EdgeConnectorTest.java
